+++ spotbugs/src/main/java/edu/umd/cs/findbugs/filter/SourceMatcher.java
@@ -28,6 +28,7 @@ import org.slf4j.LoggerFactory;
 
 import edu.umd.cs.findbugs.BugInstance;
 import edu.umd.cs.findbugs.ClassAnnotation;
+import edu.umd.cs.findbugs.SystemProperties;
 import edu.umd.cs.findbugs.xml.XMLAttributeList;
 import edu.umd.cs.findbugs.xml.XMLOutput;
 
@@ -47,15 +48,30 @@ public class SourceMatcher implements Matcher {
 
     @Override
     public boolean match(BugInstance bugInstance) {
+        // Match on the pure filename first
         ClassAnnotation primaryClassAnnotation = bugInstance.getPrimaryClass();
         if(primaryClassAnnotation == null){
             return false;
         }
+        
+        // Create the variable to store the result. If there is no valid bug file name,
+        // the result is false no matter what.
+        boolean result = false;
         String bugFileName = primaryClassAnnotation.getSourceFileName();
-        if(bugFileName == null || bugFileName.isEmpty()){
-            return false;
+        if(bugFileName != null && !bugFileName.isEmpty()) {
+            
+            // Check if the files are already matching and store the result
+            // This is the check which just compares in a simple way and does not obey
+            // the full path of the file.
+            result = fileName.match(bugFileName);
+            
+            // if no result try again to match with the full path as well
+            if (!result && bugInstance.getPrimarySourceLineAnnotation().isSourceFileKnown()) {
+                bugFileName = bugInstance.getPrimarySourceLineAnnotation().getRealSourcePath();
+                result = fileName.match(bugFileName);
+            }
         }
-        boolean result = fileName.match(bugFileName);
+        
         LOG.debug("Matching {} with {}, result = {}", bugFileName, fileName, result);
         return result;
     }